Fix ambiguous internal library linkage mode
The add_library CMake lines for several internal libraries are ambiguous. This
can cause problems when certain systems are configured for dynamic linkage
compilation mode (e.g. Gentoo Linux defaults to "BUILD_SHARE_LIBS=ON"),
causing the internal libraries to be compiled but not linked to the ecwolf
executable as expected.

This patch fixes the ambiguity by specifying a STATIC mode for internal
libraries, thus resolving the build issues experienced on Gentoo Linux and
other similar system environments.