Reorganize serializing of int's
On 32-bit systems both int and long are 32-bit wide. Depending
on platform int32 and SDWORD may be defined through either of them. Same
problem with several other types. We tried to make sense out of it but
it varies between platforms drastically. Compiler refuses to compile
if either same basic C type is handled twice or we miss a type that
we actually use. Given that we serialize int's, int32's and SDWORD's, the
simplest solution is to write a serializer for all 11 basic integer C types.

This patch implements this approach